campus-admin-agent-final/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
sse-starlette==1.6.5
groq==0.5.0
aiofiles==23.2.0